+++ gems/factory_bot-blueprint-rspec/lib/factory_bot/blueprint/rspec.rb
@@ -2,4 +2,5 @@
 
 require "factory_bot/blueprint"
 require_relative "rspec/version"
+require_relative "rspec/letbp"
 require_relative "rspec/driver"