+++ gems/factory_bot-blueprint-rspec/lib/factory_bot/blueprint/rspec/driver.rb
@@ -4,7 +4,7 @@ module FactoryBot
   module Blueprint
     module RSpec
       # Helper methods to integrate <code>factory_bot-blueprint</code> with RSpec.
-      # This module is automatically extended to {RSpec::Core::ExampleGroup}.
+      # This module is automatically extended to <code>RSpec::Core::ExampleGroup</code>.
       #
       # To use FactoryBot::Blueprint from RSpec with minimal effort, usually {#letbp} is the best choice.
       module Driver
@@ -30,8 +30,14 @@ module FactoryBot
           name
         end
 
-        # Build objects by <code>build</code> build strategy in FactoryBot from a blueprint and declare them using
-        # RSpec's <code>let</code>.
+        # <code>let!</code> version of {#let_blueprint}.
+        def let_blueprint!(...)
+          name = let_blueprint(...)
+          before { __send__(name) }
+        end
+
+        # Create a set of objects by <code>build</code> build strategy in FactoryBot from a blueprint and declare them
+        # using RSpec's <code>let</code>.
         # @param map [Hash{Symbol => Object}]
         #   map data structure from source blueprints to instance definitions.
         #   Each instance will be built with <code>FactoryBot::Blueprint.build(__send__(source))</code>
@@ -42,18 +48,18 @@ module FactoryBot
         #     end
         #
         #     # Simplest example:
-        #     # This is equivalent to `let_blueprint_build blog_bp: { result: :blog }`
-        #     let_blueprint_build blog_bp: :blog
+        #     # This is equivalent to `let_blueprint_build(blog_bp: { result: :blog })`
+        #     let_blueprint_build(blog_bp: :blog)
         #
         #     # Another shorthand example:
-        #     # This is equivalent to `let_blueprint_build blog_bp: { items: %i[blog article] }`
-        #     let_blueprint_build blog_bp: %i[blog article]
+        #     # This is equivalent to `let_blueprint_build(blog_bp: { items: %i[blog article] })`
+        #     let_blueprint_build(blog_bp: %i[blog article])
         #
         #     # Most flexible example:
         #     #   :result specifies the name of the result object to be declared. Defaults to nil
         #     #   :items specifies the names of the objects to be declared. Defaults to []
         #     #   :instance specifies the name of the instance object to be declared. Defaults to :"#{source}_instance"
-        #     let_blueprint_build blog_bp: { result: :blog, items: %i[article], instance: :blog_instance }
+        #     let_blueprint_build(blog_bp: { result: :blog, items: %i[article], instance: :blog_instance })
         #
         #     # Above example will be expanded to:
         #     let(:blog_instance) { ::FactoryBot::Blueprint.build(blog_bp) }       # the instance object
@@ -62,21 +68,30 @@ module FactoryBot
         #   end
         def let_blueprint_build(**map) = let_blueprint_instantiate(:build, **map)
 
-        # Build objects by <code>build_stubbed</code> build strategy in FactoryBot from a blueprint and declare them
-        # using RSpec's <code>let</code>.
+        # <code>let!</code> version of {#let_blueprint_build}.
+        def let_blueprint_build!(**map) = let_blueprint_instantiate!(:build, **map)
+
+        # Create a set of objects by <code>build_stubbed</code> build strategy in FactoryBot from a blueprint and
+        # declare them using RSpec's <code>let</code>.
         # See {#let_blueprint_build} for more details.
         # @param map [Hash{Symbol => Object}]
         #   map data structure from source blueprints to instance definitions.
         #   Each instance will be built with <code>FactoryBot::Blueprint.build_stubbed(__send__(source))</code>
         def let_blueprint_build_stubbed(**map) = let_blueprint_instantiate(:build_stubbed, **map)
 
-        # Build objects by <code>create</code> build strategy in FactoryBot from a blueprint and declare them using
-        # RSpec's <code>let</code>.
+        # <code>let!</code> version of {#let_blueprint_build_stubbed}.
+        def let_blueprint_build_stubbed!(**map) = let_blueprint_instantiate!(:build_stubbed, **map)
+
+        # Create a set of objects by <code>create</code> build strategy in FactoryBot from a blueprint and declare them
+        # using RSpec's <code>let</code>.
         # See {#let_blueprint_build} for more details.
         # @param map [Hash{Symbol => Object}]
         #   map data structure from source blueprints to instance definitions.
         #   Each instance will be built with <code>FactoryBot::Blueprint.create(__send__(source))</code>
         def let_blueprint_create(**map) = let_blueprint_instantiate(:create, **map)
+
+        # <code>let!</code> version of {#let_blueprint_create}.
+        def let_blueprint_create!(**map) = let_blueprint_instantiate!(:create, **map)
 
         # @!visibility private
         def let_blueprint_instantiate(build_strategy, **map)
@@ -121,22 +136,25 @@ module FactoryBot
           end
         end
 
-        # Write the blueprint in DSL, create an instance of it, and declare each object of the instance using RSpec's
+        # @!visibility private
+        def let_blueprint_instantiate!(...)
+          names = let_blueprint_instantiate(...)
+          before { names.each { __send__(_1) } }
+        end
+
+        # Write the blueprint in DSL, create a set of objects from it, and declare each object using RSpec's
         # <code>let</code>.
         #
-        # This is a shorthand for {#let_blueprint} with {#let_blueprint_build} or {#let_blueprint_create}.
+        # This is a shorthand for a combination of {#let_blueprint} with {#let_blueprint_build}
+        # (or other build strategy).
         # @param name [Symbol]
         #   name of the result object to be declared using RSpec's <code>let</code>.
         #   It is also used as a name prefix of the blueprint
         # @param items [Array<Symbol>] names of the objects to be declared using RSpec's <code>let</code>
-        # @param inherit [Boolean] whether to extend the blueprint by <code>super()</code>
-        # @param strategy [:build, :build_stubbed, :create]
-        #   FactoryBot build strategy to use when building objects.
-        #   This option is ignored if <code>inherit: true</code>
-        # @yield Write Blueprint DSL code here
+        # @return [Letbp]
         # @example
         #   RSpec.describe "something" do
-        #     letbp(:blog, %i[article]) do
+        #     letbp(:blog, %i[article]).build do
         #       blog(title: "Daily log") do
         #         let.article(title: "Article 1")
         #         article(title: "Article 2")
@@ -152,47 +170,15 @@ module FactoryBot
         #         article(title: "Article 3")
         #       end
         #     end
-        #     let_blueprint_create blog_blueprint: { result: :blog, items: %i[article] }
+        #     let_blueprint_build(blog_blueprint: { result: :blog, items: %i[article] })
         #   end
-        def letbp(name, items = [], inherit: false, strategy: :create, &)
-          raise TypeError, "name must be a Symbol" unless name.is_a?(Symbol)
-
-          source = :"#{name}_blueprint"
-          strategy = nil if inherit
-
-          let_blueprint(source, inherit:, &)
-          let_blueprint_instantiate(strategy, source => { result: name, items: })
-        end
-
-        # <code>let!</code> version of {#let_blueprint}.
-        def let_blueprint!(...)
-          name = let_blueprint(...)
-          before { __send__(name) }
-        end
-
-        # <code>let!</code> version of {#let_blueprint_build}.
-        def let_blueprint_build!(...)
-          names = let_blueprint_build(...)
-          before { names.each { __send__(_1) } }
-        end
-
-        # <code>let!</code> version of {#let_blueprint_build_stubbed}.
-        def let_blueprint_build_stubbed!(...)
-          names = let_blueprint_build_stubbed(...)
-          before { names.each { __send__(_1) } }
-        end
-
-        # <code>let!</code> version of {#let_blueprint_create}.
-        def let_blueprint_create!(...)
-          names = let_blueprint_create(...)
-          before { names.each { __send__(_1) } }
-        end
+        def letbp(name, items = []) = Letbp.new(self, name, items)
 
         # <code>let!</code> version of {#letbp}.
-        def letbp!(...)
-          names = letbp(...)
-          before { names.each { __send__(_1) } }
-        end
+        # @param name [Symbol]
+        # @param items [Array<Symbol>]
+        # @return [Letbp]
+        def letbp!(name, items = []) = Letbp.new(self, name, items, eager: true)
       end
     end
   end