+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added: Add `build_stubbed` build strategy support
 - Added: Add `FactoryBot::Blueprint::Methods`
+- **Changed**: `letbp` syntax changed to `letbp(...).<build_strategy>`
 
 ## [0.4.0] - 2024/08/13